requirements.txt
Flask
flask_login
flask_sqlalchemy
flask_inject
python_dotenv
werkzeug